json-generator/requirements.txt
# Core dependencies
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5
openai==1.10.0
ollama==0.1.7
llama-cpp-python==0.2.32
transformers==4.37.0
torch==2.7.1

# API Framework
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.0.0

# Utilities
aiohttp==3.9.1
requests==2.31.0
tqdm==4.66.1
colorama==0.4.6
rich==13.7.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Development tools
black==23.12.1
ruff==0.1.11
pre-commit==3.6.0
